requirements.txt
fastapi==0.115.6
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
transformers==4.48.0
datasets==3.2.0
lightgbm==4.5.0
catboost==1.2.7
pyyaml==6.0.2
scipy==1.15.0
tokenizers==0.21.0
uvicorn==0.32.1
tqdm==4.67.1